requirement.txt
asgiref==3.8.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==4.2
django-filter==24.3
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.2
flake8==7.1.1
idna==3.10
importlib_resources==6.4.3
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
